Add Command enum
Since there are only a fixed set of commands, we can use an enum to
represent the commands.

Let's change all command constants to values under the Command
enum. User input is now trimmed and a static contains method is
added to Command enum.

This allows us to convert user input from type String to Command.