Modify user query to be split into command and arguments
User query is a single string. This makes it difficult to
create commands with custom arguments, such as the `mark` command
from Level-3.

As such, let's divide the user query into two strings,
a "command" string and an "argument" string, split
by whitespace of some length.

Splitting it in this manner keeps the syntax familiar to users
familiar with the terminal, whilst maintaining the ability to
have whitespaces for arguments without dealing with escape
characters.