Add automated testing
There is currently no automated testing in-place.

We want to ensure the Poolsheen program runs as expected without
needing to manually test every time a change is made.

Let's,
* create an input and expected file
* create a runtest.bat file which can be executed to run the tests

The INPUT.txt will be represent the inputs into the program.
The EXPECTED.TXT will be what the program should print and will be
compared with ACTUAL.TXT, the actual lines that our program will
print.