Add error handling
There is currently no handling of edge cases when wrong
or incomplete inputs are entered.

We want Poolsheen to prompt the user on how they can fix their
wrong or incomplete inputs if an error occurs.

Let's,
*create an IncompleteCommandException class for incomplete commands
*create an UnknownCommandException class for unknown commands
*catch runtime errors with IndexOutOfBoundsException
and NumberFormatException
*catch unexpected errors using Exception

The IncompleteCommandException when caught will let the user know
what part of the command they are missing. The
UnknownCommandException lets the user know that the command does not
currently exist.