Complete Level 4. ToDos, Events, Deadlines
addTask now takes in Task instead of String.

This is to pass the logic of deciding what kind of a task to the main
logic loop. This is due to the different fields a task may have. Making
addTask handle such decisions can lead to many parameters needed and
hard to maintain.

Messages are now divided into start, content and end.

We now require intro messages as well as end messages for any command
input. Splitting message up will allow us to have greater readability.