Add delete task
There is currently no way to remove tasks that a user has
accidentally added into their Poolsheen program.

We want Poolsheen to include a "delete" command to delete a task
of position x from the list of tasks that Poolsheen remembers.

Let's create a new method called deleteTask which removes the task
of position x.

The position x will be fed after the input has been parsed. If
"delete 5" is inputted and the list of tasks has a size greater
than 5, the task of position 5 will be removed.