Add enum types for commands
Commands are now only distinguished by the string commands that
they are assigned to.

We want each command to have a type using the Java enum type.

Let's create a new enum type for each command type.

The switch cases for processing the input will now compare using
the value of the string which will be converted to an enum type if
that enum type exists.