Duke.java: Refactor main loop if-else into switch
Using if-else statements for main loop had bad readability. This is due to the denseness of the statements. Using switch statement will allow for better readability since the conditions are clear. One trade off is that the "bye" condition cannot exist within the switch statement due to the location of break. However, readability is more important and the "bye" condition is close enough to the switch statements.