Complete A-Enums
Command.java: getIfPresent
This is implemented in order to handle any inputs that may be invalid.
This is because by default, Enum.valueOf throws an
IllegalArgumentException upon an invalid argument.

As it is likely that the input can be invalid and this is more of an
checked exception, let's use getIfPresent to handle wrong user inputs
and return Command.INVALID to signify wrong user inputs.