Duke.java: Move runCommand into its own method
The entry point of Duke.java is overcrowded as command handling code are
all written in it. By moving runCommand into its own method, we are able
to better compartmentalize our code. We can move the main application
loop into its own method as well. This allows our entry point to simply
be the driver.