Add file I/O for persistent state
We need a common interchange format for Tasks to interact with
files. We also need a way to extend the testing to deal with the
case where the file is missing.

Let's add common String[] method getAsStringArray() format for
Tasks. To make it easier to make the tasks, let's also add
TaskFactory to utilise the Factory design pattern.

Let's come up with a new encoding for the file. As Base64 methods
are available in Java 11, this can be done by encoding string
information for each file in Base64, and separated with commas on
a single line.

Let's run the test twice, once after deleting the duke.txt file and
once after not deleting the file. This would allow us to do
regression testing in both cases.