Level 7: Save
The chatbot allows users to add and modify the Tasks List but does not allow these tasks to be saved and used again.

Let's update the chatbot to allow the saving of tasks into an output file, duke.txt and allow the data from this file to be loaded when Duke starts up.