Add a save file behaviour
To do list created by user is not saved to any file.

The creating of a text output allows the user to make use of the to-do list they have created using our program.

Let's make use of the File and FileWriter classes to create and write to a file that the user can access.

The file can be accessed under iP/data/duke.txt