Make the code more OOP
Refactor the code to extract out closely related code as classes.
Ui: deals with interactions with the user.
FileIO: deals with loading tasks from the file and saving tasks in the file.
Parser: deals with making sense of the user command.
StorageList: contains the task list e.g., it has operations to add/delete tasks in the list.
Command: contains the abstract logic for *Command classes that inheirits from it.

By refactoring the code to be more OOP, we are able to model
standard working modules to communicate with one another, this in turns allows us to
break the program into the bit-sized problems that can be solved easily. This will give us
greater programmer productivity, better quality of software and lesser maintenance cost.
This will also means that we can eliminate redundant code and extend the use of existing classes,
by using inheritance. However, this will definitely increase the size of the code base.