Add Date Functionality
To do list takes any string, and implements it as the "deadline" for activities.

Accepting and displaying date of activities would improve the implementation of Duke.

Let's make use of Java's LocalDate and DateTimeFormatter classes to read and write dates in specific formats.

Using Java's in-built classes allow us to store the Date as objects that can be accessed and reformatted