Add support for Dates and Times in Duke Bot
Modify Deadline and Event to make use of LocalDateTime objects to store the Dates and Times
Change the representation of Event objects to store a start and end LocalDateTime object to use
for storing LocalDateTimes more easily and changed it's constructor. Modify Duke bot to read in
the inputs correctly.