Add Map taskOperations
This map contains the functions that specify how to deal with the
various user commands within Duke.

The code is initially lengthened by the use of if else statements to
check for various commands. Now they are replaced with a single
encompassing if statement that retrieves the required function to handle
user request.