Complete Level 7. Save
Constructor no longer initializes a TaskList. It is more natural for
startDuke() to initialize a TaskList as a TaskList is only needed when
Duke is started. Furthermore, this allows us to add the logic to read
from storage in a cleaner fashion (inside startDuke).

Other changes:
addTask, deleteTask, markTask and unmarkTask will now make changes to
the persistent storage file using Storage's static methods.