Implement Level-7
This commit will add FileManager class so as to handle the read, create and writing functionality of the file.

It has also made some minor modifications to the Task, ToDo, Deadline and Event class, adding the toFileString()
method, which prints out the data in a format to be stored in the .txt file.

The Duke class has also been modified to incorporate the writeToFile functionality whenever there are changes.