Enable Duke to process inputs by users for time-related inputs to be processed as LocalDateTime for Deadline and Event task that has time-related attributes
Duke can only process time-related attributes (due for deadline and duration for event) as String

This limits the manipulation of these field and makes it difficult for printing the inputs in a different format.

Hence, we should add this functionality to improve the versality of Duke such as to enable it to print out inputs in a different format.