Add function to parse dates and convert between date formats
Increment Level-8 requires us to implement a way for Duke to handle
dates. We can go further and enable Duke to handle times as well.

Let's implement a way for Duke to parse, store and format convert
dates.

For now, handling times is not implemented but will try to in the
future.