Add more OOP
Currently most of the code resides in the Poolsheen.java file.

As the Poolsheen program expands, it becomes harder to maintain if
everything is kept inside one file. So, we can split what Poolsheen
can do into 4 classes.

Let's,
- create a Ui, Storage, Parser, and, TaskList class
- create an abstract Command class and its subclasses

We can now test the Poolsheen functionality based on each class
rather than the whole class. It is easier to change each command
as well.