Save tasks in the hard disk.
The task list resets every time the program is run.

Tasks added to the task list are not saved. The task is is blank whenever the program is started.

Let's save the tasks in the hard disk automatically whenever the task list changes and load the data from the hard disk when Duke starts up.