Add JUnit tests
Currently, there is only automated UI text testing.

More forms of automated testing need to be included to ensure the
maintainability of the program.

Let's create unit tests for most classes of the Poolsheen program.

Unit tests allow developers of the Poolsheen program to isolate
bugs to their respective classes, making it easier to debug code.