Add Parser class
Duke class was responsible for processing client inputs and categorising the tasks into the different types for creation.

It would be better to create a class specific for this purpose, letting the Duke class mostly only call on other classes and to do minimal conditionals.

The parser class is addded to process user inputs and to categorise the commands for different Duke functionality.