Add the ability to find tasks
There is no way to look through the to-do list to find tasks.

Providing a way for users to look for existing tasks would be a useful upgrade.

Let's add a method as well as a case under the parser that supports the find command.

This increases the capability of our parser, and makes good use of existing classes.