Add Increments as parallel branches
Currently, the program lacks JavaDoc comments, and does not have a way to find specific tasks.

The prgram should have JavaDoc comments and a way to find tasks.

Let's add JavaDoc comments, tweak the code to comply with a coding standard and give users a way to find a task by searching for a keyword.