Add task to build.gradle file to automate building of fat jar and remove method not used from Ui class.
askForPath is a method from the Ui class which is not used. There is no task to create a fat jar via build.gradle.

askForPath affects readability and adds extra lines of unnecessary codes. We also need a task to create a fat jar via build.gradle.

We should remove askForPath method sicne it is nto used and add task to build.gradle in order to automate the creation of fat jar file.