Reorganises invocation flow of command classes
The last commit had commands with functions that do not serve concrete purposes. This commit ensures that execute() is used to perform looping instructions until termination and parse(String) is used to parse new commands that will be interpreted within this class, either from the execute method, or from an external user that calls parse on this object.