Remove method add from Storage class, implement add method in TaskList instead. Added Java documentation.
add method is in Storage class but it deals with adding tasks into the task list. There is no Java documentation.

TaskList class should be handling this operaton. Java documentations should also be added for the classes.

Transfer the add method to the TaskList class from the Storage class. Added Java documentation for the classes of the program.