Add test cases for Level-8
The test files input.txt and EXPECTED.TXT do not contain
test inputs to test functions implemented in Level 8 (Dates and Times).

Tests should be included to test the behaviour of program
for Level-8 and also in future modifications.

Let's update input.txt and EXPECTED.TXT to include those tests.