fix: getTaskIndex() should return 0-based index
output of getTaskIndex() is used directly in UpdateCommand subclasses
get and mark/unmark a task.

however, the user passes in a 1-based index,
but the taskList methods takes in 0-based indexes.

thus it is more accurate for getTaskIndex() to return a 0-based index,
that can then be used directly in taskList methods.