refactor: storage class
instead of in Duke class, initialise FileWriter class, write() and
close() in Storage class instead, since Storage should take care of all
interactions with the output file.

write() to be used to write changes after the TaskList is changed by a
command.

list tasks with ListCommand upon loading of tasks in load() completes.