refactor: write changes to taskList into output file
in TaskList class, every method that results in change to the tasks will
now write to the output file, by calling storage.write(). thus, the
output file is updated whenever there are changes to the tasks due to a
command. when catching the IOException, changes made to tasks and
addCommands are reversed, since we are unable to save this change to the
output file, then the IOException is rethrown with a message.

in Command class, execute() should take in storage, which will be passed
into command.execute(), then to the methods in TaskList. execute() will
only throw IOException if no Ui is passed in, as there is no way to
print a message to inform the user that an error has occured.

in TaskList class, mark() and unmark() added, which were initially only
in Task class. this is these cause an update to the TaskList, which
should be updated with storage.write(). in toString(), streams are used
instead of concatenating, since it is more efficient.