fix: catch exceptions in Duke class
IllegalArgumentException thrown by Parser.parse should be caught and
their messages printed, as this denotes a wrong command being entered.
otherwise, this would cause the program to be terminated upon a wrong
command.

catch IOException now thrown by storage.load(), and assign a new empty
instance of TaskList. this is to prevent a half-loaded TaskList from
being used, in the event that the IOException was thrown when some lines
had already been read from the output file.