Add Comments to all public methods and classes
Many public classes and method did not originally have comments to
explain/describe its functions.

Ensuring that comments are added to each public method and class
would greatly increase code readability and is in accordance to the
Java Coding Standards prescribed.

The addition of comments would make it much easier for other people
to understand the respective uses of the classes/functions.