feat: find command
execute() in FindCommand uses a new method search() in TaskList, which
returns a TaskList instance representing the matching tasks. search()
filters the tasks and addCommands with the condition that the
corresponding task's text field contains the keyword. search()
uses a new private TaskList constructor that takes in predefined tasks
and addCommands lists.