Add help functionality
Users can know how to use the program with the help functionality.
Initially, there were no guidance on which command to type, and the
input format of each of the commands. Users needed to reply on memory.
Now, users can have a quick reference in-app to find out the command
they want.

The help functionality is added in two places:
- As a button at the bottom right corner
- As a command 'help'