Add checkstyle
Currently, there are not many tools that can detect coding style
violations.

This needs to change as coding style standard and quality is
important for maintaining the project.

Let's add the Checkstyle tool.

The Checkstyle tool is a static analysis tool that can detect
errors in code before compilation. It prevents errors early.

Refer to this link for how the Checkstyle tool is implemented
https://se-education.org/guides/tutorials/checkstyle.html#configuring-checkstyle