Fix markTask, unmarkTask and deleteTask deleting the wrong task.
Previously, the methods take in the input as the index.
It is changed to take in the input as the numbered task in the list,
then it will internally adjust to the index to delete.

For example, to delete the first task in the list,
in the previous version it would be delete 0, now it will be delete 1.