Improve code quality
Previously, certain methods that perform the same task have different
parameter names. As such, parameter names are changed to be more
consistent in general. For example, delete() has int i as the method
parameter, the same way as markTask() and unmarkTask(), while it has
int index as the method parameter previously.

This would improve consistency across the code base.