Add some assertions to Duke and Parser
Code currently contains no assertions.

It needs to change as using assertions is one of the project requirements and some of the assertions ensure assumptions that parameters are non-null might be useful.

Adding these assertions let's us check that we have handled the null cases for filename and provided a non-null consumer to Duke.