Change DateTimeFormatter as a static attribute in Task
The byDateTime and atDateTime attributes in Deadline and Event,
respectively, have a common usage DateTimeFormatter, which is
to output the LocalDateTime attribute as a formatted String.

This is an instance of code duplication.

Let's put the result of the DateTimeFormatter.ofPattern method in
Task as a static final attribute called OUTPUT_DATETIME_PATTERN.

This will not only reduce code duplication, but also allow other
subclasses of Task to handle datetime-related features easier in
the future. This includes but not limited to:
* storing the datetime the task is marked or unmarked
* storing the datetime the task is created or modified
* checking for datetime validity