[A-FullCommitMessage] Change datetime pattern for output
The datetime pattern for outputting Task objects as a String can be
ambiguous at times.

The year and time of the event can have the exact same value, for
example, 20:22 at the year 2022.

By adding a colon between the hours and the minutes, there should be
a clear distinction between the year and the time of a datetime string.

Let's add a single colon on the static final attribute called
OUTPUT_DATETIME_PATTERN and wrap to a new line to avoid possible
CheckStyle errors.