Use Assertions in TaskList methods
There are several methods that may fail if the parameters provided to
them are invalid.

Using Assertions to define assumptions about the program state enables
the runtime to verify them. This helps to detect mistakes in the code.

Let's document important assertions that should hold at various points
in the code.