[A-Assertions] Use Assertions
Assert statements can be used at some parts of the code to double
check on trivial things such as non-negative task ID, and using regex
to capture the strings on each line of an existing task data file.

Let's put some basic assert statements that might be useful when giving
out-of-the-box inputs as an user, this includes negative task ID input
and modifying the local task data file manually.