Add undo command support
There is no way to undo a command.

This might inconvenience the user if they made an unintentional command
or wants to revert their last command.

Let's add support for undo command. The undo command is executed if the
user enters `undo`.

The undo command will revert the last command which modifies the task
list. Note that the undo command can only undo a command which was made
since Duke is launched (i.e., there is no way to undo a command made in
a previous launch of Duke). Additionally, there is no way to undo
an undo command.