Add new custom Duke exception called FileCorruptedException
There is no existing exception handling for cases where the storage file
containing the tasks are modified and cannnot be parsed properly.

Adding an exception that handles such a case allows the program to
continue running with an empty task list. Also, it allows for better
if-else code structure.

Let's create a new DukeException, FileCorruptedException, which extends
DukeException.