Implement mass operations
The delete, mark and unmark commands all take 1 task index as an
argument.

Performing the same operation on multiple items requires many commands,
and is slow.

Adding the ability to specify multiple task indices as as the arguments
for delete, mark and unmark commands will allow the user to use the app
more efficiently.