Add EditDescCommand to update description of Tasks
Users have no way of editing existing Tasks.

Users may want to edit Tasks without deleting and creating a new
one.

Let's add an EditDescCommand to update the description of an
existing Task

Users may want to edit only a certain property of a Task

EditTimeCommand and EditStatusCommand will be added in the future