Add assertions
Add assertions to check for assumptions at different parts of the program.

In the parseUserInput method, the switch case block should never reach the default case.

The program should either throw a InvalidCommandException or it will be one of the keywords.

In the method to read and process the input file, we need to make sure the file exists in the first place.