Add assertions when adding new tasks
When the `add`, `addFromFile` and `getDescription` methods are called,
the input is assumed to have been a `deadline`, `event` or `todo`
command. Other commands are ignored (without warnings or exceptions).

Errorneous behaviour that results in calling `add`, `addFromFile` or
`getDescription` on inputs that are not `deadline`, `event` or `todo`
tasks can be picked up by adding assertions. This will aid in future
debugging.

Let's add an `assert` statement in the `default` and `else` blocks of
`add`, `addFromFile` and `getDescription.