Add streams to tasklist methods
Find and Due commands are done by iterating through for loops.

To allow these commands to be more flexible in the future for sorting
etc.

Utilize streams for their filter method and collect them into a list.

Easier readability as well as being able to add more methods to the
stream if needed.