Add and enable Assertions
Currently, important assumptions are checked using if statements
or not checked at all.

This is dangerous as it is not guaranteed that the assumptions will
hold. Additionally, there is a risk that some assumptions might be
forgotten about.

Now, all important assumptions are checked using Assertions.

Assertions allow assumptions to be clearly marked out from other code.