GUI: transition to a GUI
Using the CLI-mode is difficult and not visually appealing.

Responses in the CLI mode are also cumbersome to add, with delimitations made using lines and indentations.

Let's use JavaFX to add a GUI. Let's also simplify the generation of Duke's responses in the code.

Using JavaFX is recommended because JavaFX is a versatile and popular library to create a desktop GUI. The resulting UI is simple and minimalistic.