Add Javadoc comments and modify Stashy-Ui interaction
Some of the Javadoc comments on non-void methods have no @return tag.

Ui should have a method that shows the error message as a String instead
of a void. This way, the additional text can be shown in the GUI as well
if there is any StashyException caught during the command execution.