Replace use of DukeException with CommandResult
DukeException serves no real functional purpose with the creation of
CommandResult. It only messes up the code and makes it harder to
navigate and untidy. CommandResult can already encapsulate messages,
including both success and failure messages, to be sent to the users.
Since exceptions are used to handle user fault, it is okay for the user
to see the exception (which actually complies with previous behaviour).

Let's remove the use of DukeException and wrap error messages that the
user is supposed to see anyway within the CommandResult.