Add Assertions to TaskList and Storage
Use `assert` feature to document important assumptions that should
hold at various points in the code.

Assertions added to include checking of the right command-word and
the length of the command String array in the functions of TaskList.

Assertions added to ensure the data stored in the save file is valid
before loading it into Duke.