Standardize error messages and improve documentation
Currently error messages are not very standardized.

The console needs to be used for useful messages. The console can get
messy when adding more error messages to the console if error messages
are not standardized.

Let's standardize the error message by ending it with "- terminating"
and throwing a RuntimeException which is not caught so that it can
fail quickly.