Add assertions to Duke and Task class
There are no assertions in Duke and Task classes.

There are some assumptions present that the code in said classes
use while running (E.g getResponse() in Duke assumes a non-null
UI reference in its class).

Adding assert statements ensure those references are non-null and
truly present which can help in debugging code.

Let's add assert statements to check that references are non null
in Duke and Task classes.

Using assertions is preferable over if statements as it conveys
that the code is used for debugging purposes and makes the code
look cleaner.