DukeException: collate all exceptions thrown
Restructure all the exceptions in the Command sub-classes (e.g.
AddCommand, FindCommand, MarkCommand, etc.) to make the code more
understandable.

Converting all the exceptions to DukeException with a message to be
display later on.

Let's catch all the exceptions appear then throw a new DukeException
with a new message embed. The DukeException is then handled on another
class.