Improve code quality in UI class.
There are several methods in UI class that contains magic strings.

Magic strings is not good for extensibility as it is harder for
people to understand its purpose.

Lets use named constants to denote such strings and return them.

This way it will be more easily understood by people reading the
code.