Enable GUI to display longer messages in full
When Duke's message is too long, the text gets shortened with ellipsis.

Users are unable to access their full task list when there are too many
items, limiting the usefulness of Duke's task list system.

Let's adjust height limitations for the Label and HBox in the DialogBox
to follow the preferred height needed to fit the whole message.