Add assertion to check if task is defineed when decoding
The tasks in the .txt file should start with either T,E or D.
There might be a scenario where the .txt file is modified by
the user, resulting in this to not be the case.

Let's write an assertion to assert that the task is not null
at that point.