Add assert
Assert is not added in this project

Assert should be added so that the programmer can spot his or her errors more easily

Added some asserts in the dukepro.handlers.Decoder and dukepro.handlers.Interact.

The aforementioned classes are responsible for handling user input, and throwing errors when users provide bad input. The programmer may handle errors and parse inputs wrongly so these classes need assert to be added