Use Java assertions
There are important assumptions at various points in the code.

The use of assertions can verify if these assumptions hold true and
can help detect potential bugs.

Let's add assertions in some of the methods where the argument can be
null but should not be invoked with a null argument at any point.