Improve Parser class code quality.
parseInput() method is very long with deep nesting. It also
has many level of abstraction, violating the SLAP principle.
It is difficult to understand the code. There are also magic
literals present.

Let's apply the SLAP principle and break the parseInput() method
into multiple smaller methods, as well as refactor code to reduce
the level of nesting.

It would be easier to read and understand the code in the future,
as well as making it easier to debug.