Add assertions
No assertions is being used to verify program state.

If program is not in assumed state due to bugs, running the program
may cause problems. The sooner the execution is stopped, the safer
it is.

Assertions are added to TaskList to verify date format and task
format of data loaded from save file.