Duke::start initialises and sets up various JavaFx objects
The current implementation of the method is too long (>30 LoC) and has more than a single level of abstraction.

Let's abstract out similar functionalities into private helper methods enables us to achieve a single level of abstraction and greatly shorten the length of the code within the method.

This will enhance the overall readability of the codebase.