Add assertions in Command classes
DeleteCommand, FindCommand, MarkCommand UnmarkCommand classes
updates a Task/queries for a List of Tasks.

There may be a possibility that it fails to update/query and
return null.

Assertions are added to ensure that updates/queries are done
successfully and non-null Task/List of Tasks is returned.

Assertions added to following classes:
* DeleteCommand
* FindCommand
* MarkCommand
* UnmarkCommand