#Improve Code Quality
Many public classes and method not only do not have the relevant
javaDocs or comments. They violate the correct Java Coding Standard
too.

Ensuring that these coding styles are checked and rectified helps
future developers comprehend code better as it improves readability.

The addition and rectification of code makes it easier for future
improvements for understanding the logic of algorithm.

Let's,
*rectify relevant violations
*structure code well and add relevant javaDocs