Update JavaDocs and general cleanup
JavaDoc methods are too short and undescriptive

The parameter tags are often missing or one-liners at best.

Other users reading the code might not have sufficient idea as to what is going on with the code.

Let's update the JavaDocs in line with recommendations from the textbook.

As well as checkstyle to provide more description, even in helper methods.

Doing so allows readers to understand the codebase better at a glance