Add assertions
There are no assertions in my code.

This needs to be changed as assertions help to define assumptions about
the program state so that the runtime can verify them. Assertions need
to be added to help increase code readability and error detection.

Let's add assetions to the following classes.
- Deadline
- Event
- Task
- ToDo
- DeleteCommand
- MarkCommand
- UnMarkCommand

Using assertions in these classes help to ensure that the commands
receive the right information.