Use streams when dealing with the list of tasks
For loop was used in iterating through the list of tasks.

While for loop is usable, it might run into issues with index out
of bounds when we try to retrieve items from a list using a for
loops.

Let's use streams to manage lists of tasks instead.

Doing so helps us reduce human error and makes code easier to
understand as well.