Restore regression testing for Gradle project
The old regression test had been removed because the program has been
converted from a command-line interface to a graphical user interface.

A lack of regression testing may introduce bugs as new features are
being added.

Let's modify the runtest.bat script and restore the methods intended for
printing to console to support console-based regression testing.

The regression testing is restored because it will be useful for testing
while adding new features to the graphical user interface. This helps to
verify that no functionality has been accidentally changed as a result
of adding in new features.