Refactor code to fit code quality standards
The code does not fit the code quality standards.

Refactoring the code and improving its quality improves code quality and
allows other to read my code more easily. It is also easier to fix bugs
when my code is of a better standard.

Let's update the code to fit the standard.