#Add extension C-Priority
Users want to be able to set priority of their tasks, so they can
see what needs to be done first.

Priority command function is added to Duke, so users can simply
command:
```
priority <priority level> /for <index of task in list to update>
```
(default priority level is 0)

As number of command increases, HiCommand is added so users can
see the list of instructions. Users can simply type
```
hi
```
to see the functionality of Duke.