Add assertions to code
There are no assertions in the code.

When there is a NullPointerException, the code could be hard to debug
because it can happen almost anywhere.

Let's add assertions to the code to aid debugging when variables
stored become null.

When there is a JavaFX invocation exception or similar, these can be
hard to debug. Occasionally, that could be caused by null values. By
catching this earlier, we would know which part of the code was not
working correctly by throwing up null values.