Add assertions to detect programming errors
Currently, there are no assertions in the code base to detect
programming errors.

This needs to change as assertions can help to detect programming
errors early and prevent the chat bot from performing undesirably.

Add assertions to detect programming errors.

Assertions are not only simple, but their impact on performance is considered low and worth the additional safety they provide.