Refactor parser method
The parser code is cluttered into one method, with the parser method
parsing the commands and also executing them, hence making the code hard
to read.

Let's refactor the parser code to better adhere to the single level of
abstraction principle.

This will make the code easier to read, as suggested by a reviewer of my
code (Josephaedan).

This commit also includes other minor code changes.