Add assertions
Parser :: parse takes in a String as input however it does not
account for the scenario of being passed a null object.

Although this should not occur, using assertions to verify that the
input is not null would grant assurance and safety.

Let's add an assertion to Parser::parse to prevent runtime exceptions.