Adds Assertions at Critical Points
To ensure the accuracy of this codebase and minimise error occurances, let us add asserts to specific points in the code. They signify the tautological assumptions at each point of execution, failing which implies that there is are unaccounted corner cases, which should not have led to that execution point.
Doing this will help us reduce the amount of time spent on future debugs by making our assumptions clear.