Add C-MassOps extension
My commands do not allow users to quickly clear completed tasks/ unmark
tasks efficiently.

This is not user friendly as users who want to clear completed tasks
have to clear them one at a time and it takes quite a lot of effort to
do so.

Let's add some new commands.
- Mass Delete Command
- Mass UnMark Command

These are the first few mass commands I added as these are the most
important features for now.