Provide a way to undo commands
Currently, there is no way to undo a command.

This needs to change as users might have misspelt their tasks or
accidentally deleted a task.

Provide a way to undo commands by using an undo command, followed by
specifying the number of steps to revert by.

This is done so that users can have the flexibility to revert their
state by multiple increments.