Use assertions in TaskList get and remove methods
The index verification is done by TaskList's get and remove methods.
This creates a confusion when reading the code in UpdateStatusCommand
and DeleteTaskCommand: why aren't we throwing an Exception if index is
not valid?

Having the index checking to Command's side would improve readability.

Let's move the index checking part to the commands and use assertions
in the TaskList instead.