Improve code quality
The happy path is not obvious in the TaskList class, the toString
and toStorageFormat in Deadline, Event, and Todo classes are long
and complex.

These above-mentioned issues reduce the code quality and make the
code harder to read, and thus difficult to maintain.

By extracting methods into variables in the Task class and its
subclasses. Shifted the guard clause in the TaskList, and indented
the edge cases.

These changes make the code more maintainable and readable, as
one can see the edge cases and variables in one glance easily.