Add assertions to commands and tasklist classes
There are no assertions within the code across all classes.

There is a higher risk of bugs when there are no assertions,
since we are not verifying necessary conditions.

Adding assertions to the Command allows the code to terminate
once a bug is detected during an assertion failure, reducing
the risk of bugs, for example, when the input position into
the mark command is smaller than 0, there is most likely a
bug in the code.